+++ src/components/common/index.ts
@@ -1,3 +1,2 @@
 export * from './Biography'
-export * from './MetaTag'
 export * from './Navigation'
