+++ src/components/page/BlogList/BlogList.page.tsx
@@ -1,6 +1,6 @@
 import React, { ComponentProps } from 'react'
-import { MetaTag } from 'src/components/common'
 import { Layout } from 'src/components/ui/Layout'
+import { MetaTag } from 'src/components/ui/MetaTag'
 import { BlogList } from './BlogList'
 
 export type BlogListPageProps = ComponentProps<typeof BlogList>
