+++ src/components/ui/MetaTag/index.ts
@@ -1,0 +1,1 @@
+export * from './MetaTag'
